requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
fastapi[all]
uvicorn
pydantic
python-dotenv
httpx
docker